@@ -2,6 +2,7 @@
   "scripts": {
     "test": "yarn storybook:build && npm-run-all -p -r storybook:start storybook:test",
     "storybook": "rimraf node_modules/.cache && storybook dev -p 9001",
+    "build-storybook": "storybook build",
     "storybook:build": "rimraf node_modules/.cache && storybook build",
     "storybook:start": "http-server -s -p 9001 storybook-static",
     "storybook:test": "test-storybook --url http://127.0.0.1:9001 --coverage",
@@ -12,41 +13,42 @@
     "unlink": "yarn unlink storybook-addon-module-mock"
   },
   "dependencies": {
-    "@emotion/react": "^11.13.0",
+    "@emotion/react": "^11.13.3",
     "@emotion/styled": "^11.13.0",
-    "@mui/icons-material": "^5.16.7",
-    "@mui/material": "^5.16.7",
-    "@mui/styled-engine": "^5.16.6",
-    "next": "^14.2.5",
+    "@mui/icons-material": "^6.1.5",
+    "@mui/material": "^6.1.5",
+    "@mui/styled-engine": "^6.1.5",
+    "next": "^15.0.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   },
   "devDependencies": {
     "@node-libraries/scaffold": "^1.0.4",
     "@storybook/addon-coverage": "^1.0.4",
-    "@storybook/addon-essentials": "^8.2.9",
-    "@storybook/addon-interactions": "^8.2.9",
-    "@storybook/nextjs": "^8.2.9",
-    "@storybook/test": "^8.2.9",
+    "@storybook/addon-essentials": "^8.3.6",
+    "@storybook/addon-interactions": "^8.3.6",
+    "@storybook/nextjs": "^8.3.6",
+    "@storybook/test": "^8.3.6",
     "@storybook/test-runner": "^0.19.1",
-    "@types/node": "^22.3.0",
-    "@types/react": "^18.3.3",
-    "@typescript-eslint/eslint-plugin": "^8.1.0",
-    "@typescript-eslint/parser": "^8.1.0",
-    "eslint": "9.9.0",
+    "@types/node": "^22.8.2",
+    "@types/react": "^18.3.12",
+    "@typescript-eslint/eslint-plugin": "^8.12.1",
+    "@typescript-eslint/parser": "^8.12.1",
+    "chromatic": "^11.16.1",
+    "eslint": "9.13.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-import-resolver-typescript": "^3.6.1",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-storybook": "^0.8.0",
+    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-storybook": "^0.10.1",
     "http-server": "^14.1.1",
     "npm-run-all": "^4.1.5",
-    "postcss": "^8.4.41",
+    "postcss": "^8.4.47",
     "rimraf": "^6.0.1",
-    "sass": "^1.77.8",
-    "storybook": "^8.2.9",
+    "sass": "^1.80.4",
+    "storybook": "^8.3.6",
     "storybook-addon-module-mock": "^1.3.4",
-    "typescript": "^5.5.4",
-    "typescript-eslint": "^8.1.0"
+    "typescript": "^5.6.3",
+    "typescript-eslint": "^8.12.1"
   },
   "license": "MIT",
   "name": "storybook-module-mock",
